Corrigindo erro de data e refatorando parte do codigo
