+++ src/SME.SGP.WebClient/src/paginas/DiarioClasse/RegistroIndividual/componentes/DadosRegistroIndividual/novoRegistroIndividual/novoRegistroIndividualConteudo/index.js
@@ -1,0 +1,1 @@
+export { default as NovoRegistroIndividualConteudo } from './novoRegistroIndividualConteudo';
