+++ src/SME.SGP.WebClient/src/paginas/DiarioClasse/RegistroIndividual/componentes/DadosRegistroIndividual/registrosAnteriores/registrosAnterioresConteudo/index.js
@@ -1,0 +1,1 @@
+export { default as RegistrosAnterioresConteudo } from './registrosAnterioresConteudo';
